Update doc for uboot-script-gen -d
The -d option for uboot-script-gen was poorly worded.  The -d option
isn't a working directory but rather the "root" directory for where
the tftp or other root directory would be in relation to the files.  Any
files generated in the created of the boot.src/source files are in /tmp
and are deleted afterwards.

Signed-off-by: Brian Woods <brian.woods@xilinx.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>