Add options for Xen and Dom0 CMD arguments
Both XEN_CMD and DOM0_CMD are optional settings  in the config file that
set Xen's and Dom0's command line arguments.  If not set, a default
value will be used.  Also, for Dom0's command line arguments, if root
isn't set in it, it will determine the best option based on the load
argument.

Signed-off-by: Brian Woods <brian.woods@xilinx.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>