Add/fix support for SCSI/SD aliases for load cmds
With the -t option, the "sd" was an alias for the "load scsi 0:1" load
command.  Change that to a "load mmc 0:1" alias and then add a "scsi"
alias which is "load scsi 0:1".  To make sure these aliases and manually
setting the load command works, a parsing the command and setting the
correct partition and devs was also added.

Signed-off-by: Brian Woods <brian.woods@xilinx.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>