Introduce multiple partitions support to disk_image
Rootfses that should be written to disk as their own partition are
called *_ROOTFS instead of *_RAMDISK.

This change makes the config file clearer and more consistent. However
it changes the current behavior as the Dom0 RAMDISK (not ROOTFS) was
written to disk as its own partition before, and now it is not.

In the case of uboot-script-gen, dom0less guests cannot access disk
partitions yet (missing PV drivers support). Hence, if one or more domU
ROOTFSes are specified, print an error and exit. For dom0, base the
root= parameter generation on the presence of DOM0_ROOTFS.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Signed-off-by: Brian Woods <brian.woods@xilinx.com>
Reviewed-by: Brian Woods <brian.woods@xilinx.com>
(cherry picked from commit 5629a7a4a027bc78a5426bedefb3f9e803b49004)