Add support for compiling and merging partial device trees
One can specify the repository and the directory holding the partial
device trees via the config option PASSTHROUGH_DTS_REPO. This can be
either a git url or a local directory containing the partial device trees.

Also, the passthrough devices for each DOMU are specified via
DOMU_PASSTHROUGH_PATHS config option.

Imagebuilder will compile the partial device trees. The underlying
assumption is that the name of the partial device tree should
match to the name of the device specified in DOMU_PASSTHROUGH_PATHS.

For example (in case of git repo) :-
PASSTHROUGH_DTS_REPO="git@github.com:Xilinx/xen-passthrough-device-trees.git device-trees-2021.2"
DOMU_PASSTHROUGH_PATHS[0]="/axi/serial@ff010000 /axi/mmc@ff170000"

(This tells imagebuilder to clone the repo and use serial@ff010000.dts
and mmc@ff170000.dts from device-trees-2021.2 folder, to compile and generate
a partial dtb for DOMU[0].

If the files are absent, it will give an error.
Also, if the files mismatching values for #address-cells and #size-cells, it
will give an error.

Signed-off-by: Ayan Kumar Halder <ayankuma@xilinx.com>
Reviewed-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>