Avoid making changes in the device trees from the repository
One should copy the device tree to a temporary file and then make any
change (so as to enable merging of multiple device trees).
In this case, the change is to remove the marker (/dts-v1/;) from all
the device trees except the first one to be merged. DTC expects the
marker to be present only once.
Also, i, j, tmpdts and file should be set to local. This will avoid
name collision.

Signed-off-by: Ayan Kumar Halder <ayankuma@xilinx.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>