Avoid exiting the script from the common
If there is any error in compile_merge_partial_dts() and
sanity_check_partial_dts(), it should return 1 to the caller.
The caller will then check the return and call cleanup_and_return_err()
(which terminates the script).

This is to prevent compile_merge_partial_dts() and
sanity_check_partial_dts() terminating the script.
The reason being these functions can get invoked from
disk_image (in the future) which may require some additional
cleanup.

Also, moved cleanup_and_return_err() to uboot-script-gen as the
cleanup is specific to the script.

Ensure that we preseve the directory containing the partial dtbs only
when the dtbs were generated successfully. Else, we delete it.

Signed-off-by: Ayan Kumar Halder <ayankuma@xilinx.com>
Reviewed-by: Stefano Stabellini <sstabellini@kernel.org>
Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>