If bootm, generate a u-boot Legacy uImage to boot
If BOOT_CMD is specified as "bootm" it requires a uImage to boot.
Automatically generate a $XEN.ub binary using mkimage and specifying the
right loading address.

Also remove "u-boot legacy uImage" from the allowed Xen binary types
because we cannot easily check that the uImage was built with the right
address.

Note that "ARM OpenFirmware" is the type returned by certain versions of
"file" utility for a Xen arm32 binary.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
Reviewed-by: Ayan Kumar Halder <ayankuma@xilinx.com>