Add support for cache coloring
Generate the coloring information for Xen, dom0 and domUs based on
information on the config file.

Cache coloring is not upstreaming but it was published to xen-devel
already. It is very useful to be able to provide a cache coloring
configuration via config file instead of manually editing boot.source.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>