Add config option to enable PL011 emulation
Introduce DOMU_VUART_PL011 config option to enable virtual PL011 UART
for dom0less domU. If this option is not set, SBSA UART will be used
to match previous behavior.

Take the opportunity to modify type of original vpl011 property to be
bool instead of hex. Refer Xen docs/misc/arm/device-tree/booting.txt:
"An empty property to enable/disable a virtual pl011 for the guest to use"

Note that if DOMU_VUART_PL011 is not set, we use bool for vpl011 property
for backwards compatibility.

Signed-off-by: Michal Orzel <michal.orzel@amd.com>
Acked-by: Stefano Stabellini <stefano.stabellini@amd.com>